nitrous: Fix buggy uninitialized stack var use for PROC_TIMESYNC
If the kfifo is empty, the returned timestamp is supposed to be 0, since
the BT service intentionally tries to drain the kfifo upon BT turning on.

If the BT service doesn't get back a 0 for the timestamp, then it'll spin
forever on reading /proc/bluetooth/timesync, burning 100% CPU in
BluetoothCcc::OnBluetoothEnabled():
    do {
      lVar4 = getSystemTime();
    } while (lVar4 != 0);

Explicitly print out a 0 when the kfifo is empty to fix this.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>